# refine seqpool op