--- a/paddle/fluid/operators/jit/benchmark.cc
+++ b/paddle/fluid/operators/jit/benchmark.cc
@@ -195,8 +195,9 @@ void BenchSeqPoolKernel() {
   std::vector<jit::SeqPoolType> pool_types = {jit::SeqPoolType::kSum};
   for (auto type : pool_types) {
     for (int w : TestSizes()) {
+      jit::seq_pool_attr_t attr(w, type);
       for (int h : TestSizes()) {
-        const jit::seq_pool_attr_t attr(h, w, type);
+        attr.h = h;
         std::vector<T> x(h * w), y(w);
         RandomVec<T>(h * w, x.data(), -2.f, 2.f);
         const T* x_data = x.data();