… paddle/fluid/operators/jit/gen/CMakeLists.txt
… paddle/fluid/operators/jit/kernel_key.cc
… paddle/fluid/operators/jit/more/mkl/CMakeLists.txt
… paddle/fluid/operators/jit/more/mkl/mkl.cc
… paddle/fluid/operators/jit/more/mkl/mkl.h
… paddle/fluid/operators/jit/refer/CMakeLists.txt
… paddle/fluid/operators/jit/refer/refer.cc
… paddle/fluid/operators/jit/refer/refer.h
… paddle/fluid/operators/jit/test.cc
… paddle/fluid/operators/math/CMakeLists.txt
… paddle/fluid/operators/math/sequence_pooling.cc