speedup seqpool of sum pool_type.

单独比较sum seqpool， mkl 0.0343us => target 0.0062us 提升至5.5倍
```
I0107 07:23:47.070497  4942 benchmark.cc:111] Kernel Type kSeqPool: height_size[1],width_size[11],pool_type[kSum]: Refer takes 0.0275 us; JitCode takes 0.004 us; MKL takes 0.0343 us; Target takes 0.0062 us;
```

在模型中添加record比较：
math_seqpool: 1308 => 676 ms

before:
```
Event                       Calls       Total       Min.        Max.        Ave.        Ratio.
thread0::sequence_pool      770000      5333.33     0.003616    47.6168     0.0069264   0.40359
thread0::fc                 50000       3272.09     0.005852    1.69518     0.0654418   0.24761
thread0::feed               770000      1623.79     0.000932    131.12      0.00210881  0.122877
thread0::math_seqpool       770000      1308.1      0.000744    47.6041     0.00169883  0.0989881
thread0::concat             30000       522.542     0.003493    0.160611    0.0174181   0.0395425
thread0::matmul             20000       324.284     0.006234    0.57978     0.0162142   0.0245397
thread0::square             30000       224.63      0.002615    0.084313    0.00748767  0.0169985
thread0::relu               30000       204.787     0.002643    0.76185     0.00682625  0.0154969
thread0::elementwise_mul    15000       132.921     0.00463     0.093074    0.0088614   0.0100586
thread0::elementwise_sub    10000       91.5696     0.005052    0.080962    0.00915696  0.00692937
thread0::fill_constant      10000       81.9083     0.004351    0.095373    0.00819083  0.00619827
thread0::reduce_sum         5000        57.783      0.006627    0.079318    0.0115566   0.00437263
thread0::fetch              5000        36.9763     0.003942    0.077837    0.00739525  0.00279811

```

after:

```
Event                       Calls       Total       Min.        Max.        Ave.        Ratio.
thread0::sequence_pool      770000      3794.7      0.003585    38.6172     0.00492818  0.39198
thread0::fc                 50000       2670.26     0.006059    2.22126     0.0534052   0.275829
thread0::feed               770000      1340.43     0.00097     145.996     0.00174081  0.138462
thread0::math_seqpool       770000      676.867     0.000523    38.6094     0.000879048 0.069918
thread0::concat             30000       372.362     0.003385    0.150443    0.0124121   0.0384637
thread0::matmul             20000       229.95      0.006244    0.293998    0.0114975   0.0237531
thread0::square             30000       160.191     0.002561    0.060142    0.00533969  0.0165472
thread0::relu               30000       152.127     0.0027      0.178271    0.00507091  0.0157142
thread0::elementwise_mul    15000       91.2525     0.004453    0.064583    0.0060835   0.00942607
thread0::elementwise_sub    10000       66.4989     0.004996    0.036961    0.00664989  0.00686912
thread0::fill_constant      10000       58.568      0.004171    0.085623    0.0058568   0.00604988
thread0::reduce_sum         5000        40.855      0.006437    0.043902    0.00817099  0.00422018
thread0::fetch              5000        26.7998     0.004019    0.077999    0.00535997  0.00276833
```